#7 - Moved authentication functions to authentication.py
